Extract `hotReload` into its own module
