+++ self-serve.ts
@@ -4,6 +4,7 @@ import { contentType } from "jsr:@std/media-types"
 // Modules
 import * as cli from './src/cli.ts'
 import * as template from './src/templates/index.ts'
+import { generateHotReloadScript } from './src/lib/hotReload.ts'
 
 // ----
 // SELF
@@ -115,18 +116,7 @@ class Self {
         const mimeType = contentType(filePath.split('.').pop() || '') || 'application/octet-stream'
 
         if (mimeType === 'text/html; charset=utf-8') {
-            const html = new TextDecoder().decode(content)
-            const hotReloadScript = this.generateHotReloadScript()
-
-            let modifiedHtml
-            if (html.includes('</body>')) {
-                modifiedHtml = html.replace(/<\/body>/i, hotReloadScript + '\n</body>')
-            } else if (html.includes('</html>')) {
-                modifiedHtml = html.replace(/<\/html>/i, hotReloadScript + '\n</html>')
-            } else {
-                modifiedHtml = html + hotReloadScript
-            }
-
+            const modifiedHtml = this.injectHotReloadScript(content);
             return new Response(modifiedHtml, {
                 headers: {
                     "Content-Type": mimeType,
@@ -146,7 +136,21 @@ class Self {
         return new Response(content, { headers })
     }
 
+    /** Injects the hot-reload script into HTML content */
+    private injectHotReloadScript(content: Uint8Array<ArrayBuffer>) {
+        const html = new TextDecoder().decode(content)
 
+        const script = generateHotReloadScript(this.host, this.port)
+        const hotReloadScript = /* HTML */ `<script>${script}</script>`
+
+        if (html.includes('</body>')) {
+            return html.replace(/<\/body>/i, hotReloadScript + '\n</body>')
+        } else if (html.includes('</html>')) {
+            return html.replace(/<\/html>/i, hotReloadScript + '\n</html>')
+        } else {
+            return html + hotReloadScript
+        }
+    }
 
     /** Starts the file-watcher to monitor changes in the served directory */
     private async startFileWatcher() {
@@ -251,57 +255,6 @@ class Self {
                 this.wsClients.delete(client)
             }
         }
-    }
-
-    /** Generates a hot-reload script to inject into the client HTML files */
-    private generateHotReloadScript(): string {
-        return /* HTML */`
-            <script>
-                function setupHotReload() {
-                    const socket = new WebSocket('ws://${this.host}:${this.port}/__hot_reload__')
-                    socket.addEventListener('open', () => console.log('🔥 Hot-Reload WebSocket Connection Established'))
-                    socket.addEventListener('message', (event) => {
-                        const data = JSON.parse(event.data)
-                        if (data.type === 'full-reload') {
-                            window.location.reload()
-                        } else if (data.type === 'css-change') {
-                            reloadCSS(data.files)
-                        }
-                    })
-                    socket.addEventListener('close', () => console.log('Hot-Reload WebSocket Connection Closed'))
-                    socket.addEventListener('error', (error) => console.error('Hot-Reload Error: ', error))
-                }
-
-                function reloadCSS(files) {
-                    const links = document.querySelectorAll('link[rel="stylesheet"]') // Get all stylesheet links in the document
-                    links.forEach(link => {
-                        const href = link.getAttribute('href')
-                        if (!href) { return }
-
-                        // Check if this CSS file was in the changed files, and if it was, hot-swap the CSS file
-                        const shouldReload = files.some(file => {
-                            const fileName = file.split(/[\\\\/]+/g).pop() || file
-                            const linkFileName = href.split(/[\\\\/]+/g).pop() || href
-                            return linkFileName.includes(fileName) || fileName.includes(href)
-                        })
-
-                        if (shouldReload) {
-                            const newLink = link.cloneNode()
-                            const url = new URL(href, window.location.origin)
-                            url.searchParams.set('__hot_reload__', Date.now().toString())
-                            newLink.href = url.toString()
-
-                            // Replace the old link with the new one
-                            newLink.addEventListener('load', () => link.remove())
-                            newLink.addEventListener('error', () => link.remove())
-                            link.parentNode.insertBefore(newLink, link.nextSibling)
-                        }
-                    })
-                }
-
-                setupHotReload()
-            </script>
-        `
     }
 
     /** Shuts down the server and performs the necessary cleanup operation */
